change type to float

+++ src/main/java/core/basesyntax/MyHashMap.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 
 public class MyHashMap<K, V> implements MyMap<K, V> {
     private static final int DEFAULT_CAPACITY = 16;
-    private static final double LOAD_FACTOR = 0.75;
+    private static final float LOAD_FACTOR = 0.75f;
     private int threshold = (int) (DEFAULT_CAPACITY * LOAD_FACTOR);
     private Node<K, V>[] table = new Node[DEFAULT_CAPACITY];
     private int size;
